# backend/src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "LMS API Documentation",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/menus": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Menus",
                "operationId": "menus_get",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Add Menu",
                "operationId": "menus_add",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page_id",
                        "in": "query",
                        "description": "Title",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "position",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/admin/menus/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Menus",
                "operationId": "menus_delete",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit Menus",
                "operationId": "menus_edit",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Parent ID",
                        "type": "integer"
                    },
                    {
                        "name": "menu_order",
                        "in": "query",
                        "description": "Menu order",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/admin/pages": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Pages",
                "operationId": "pages_get",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID",
                        "type": "integer"
                    },
                    {
                        "name": "publish",
                        "in": "query",
                        "description": "publish",
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Add Pages",
                "operationId": "pages_add",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Tag",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Featured Image",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "publish",
                        "in": "query",
                        "description": "Publish",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/admin/pages/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Pages",
                "operationId": "pages_delete",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit Pages",
                "operationId": "pages_edit",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Tag",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Featured Image",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "publish",
                        "in": "query",
                        "description": "Publish",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/iam/login": {
            "post": {
                "tags": [
                    "IAM"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/iam/register": {
            "post": {
                "tags": [
                    "IAM"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "confirm_password",
                        "in": "query",
                        "description": "Confirm Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "confirmation_url",
                        "in": "query",
                        "description": "confirmation URL",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "Upload",
                "operationId": "upload",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "File",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}